move closable mock to mock dir from test dir
