+++ test/lib/helpers/Closable.test.ts
@@ -1,14 +1,5 @@
-import { Closable, using } from '@/lib/helpers/Closable';
-
-class ClosableMock implements Closable {
-  private closed = false;
-  getStatus = (): boolean => this.closed;
-  close = async (): Promise<void> => {
-    const promise = Promise.resolve(true);
-    await promise.then((result) => (this.closed = result));
-  };
-  test = (): Promise<string> => new Promise((resolve) => resolve('test'));
-}
+import { using } from '@/lib/helpers/Closable';
+import ClosableMock from '@/mock/Closable';
 
 describe('Closable', () => {
   test('using', async () => {
